Sloppy code of the Database
This is my first attempt at the code for the database class. I am going
to dinner with family now and can't keep working on it. I was trying to
make code to create the database if there wasn't one already, along with
all the tables if they were not there.

Side note, is this how we would want to do the add of a node to the
database? My node class is just and empty class for now.