Fold States
The node.java file in the fold States folder is what I have been working
on. I think it all works, but I have yet to test it on the way Myke has.
I wanted to do some more of that testing today if I had a chance, but
this is here in case I can't.